// web/src/api/models/TypesFriendStatus.ts
export interface TypesFriendStatus {
    /**
     * 
     * @type {boolean}
     * @memberof TypesFriendStatus
     */
    conntrackd_running?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof TypesFriendStatus
     */
    encryption: boolean;
    /**
     * 
     * @type {string}
     * @memberof TypesFriendStatus
     */
    fingerprint?: string;
    /**
     * 
     * @type {string}
     * @memberof TypesFriendStatus
     */
    hostname?: string;
    /**
     * 
     * @type {boolean}
     * @memberof TypesFriendStatus
     */
    identity_match: boolean;
    /**
     * 
     * @type {string}
     * @memberof TypesFriendStatus
     */
    last_error?: string;
    /**
     * 
     * @type {string}
     * @memberof TypesFriendStatus
     */
    last_seen?: string;
    /**
     * 
     * @type {string}
     * @memberof TypesFriendStatus
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof TypesFriendStatus
     */
    os?: string;
    /**
     * 
     * @type {boolean}
     * @memberof TypesFriendStatus
     */
    reachable: boolean;
    /**
     * 
     * @type {number}
     * @memberof TypesFriendStatus
     */
    rtt_ms?: number;
    /**
     * 
     * @type {string}
     * @memberof TypesFriendStatus
     */
    version?: string;
    /**
     * 
     * @type {Array<TypesFriendVRRPRole>}
     * @memberof TypesFriendStatus
     */
    vrrp?: Array<TypesFriendVRRPRole>;
}